Factor Photoprism API support into shared module
